+++ R/gtExtras-package.R
@@ -9,7 +9,8 @@ utils::globalVariables(
     "var", "type", "text", "cmark_rules", "rtf_wrap",
     "Var1", "rtf_escape_unicode", "value", "median", "sd",
     "name", "Mean", "SD", "n_missing", "name", "vals", "x",
-    "z_r_group_rows"
+    "z_r_group_rows","group_id", "rownum_i", "group_id", "gt_row_grps",
+    "row_id"
   )
 )