+++ R/tab_style_by_grp.R
@@ -36,14 +36,33 @@ tab_style_by_grp <- function(gt_object, column, fn, ...){
 
   stopifnot("'gt_object' must be a 'gt_tbl', have you accidentally passed raw data?" = "gt_tbl" %in% class(gt_object))
 
+  # extract group names as a character and then turn into sym
+  # for later tidyeveal
+  ## logical vector
   subset_log <- gt_object[["_boxhead"]][["type"]]=="row_group"
-  grp_col <- gt_object[["_boxhead"]][["var"]][subset_log] %>% rlang::sym()
+  ## subset vars by vector
+  grp_names <- gt_object[["_boxhead"]][["var"]][subset_log]
+  ## create a list of symbols
+  grp_col <- rlang::syms(grp_names)
 
+  # ordered levels of the row groups
+  gt_row_grps <- gt_object[["_row_groups"]]
+
+  # pull the ordered row numbers
+  grp_vec_ord <- gt_object[["_stub_df"]] %>%
+    dplyr::mutate(group_id = factor(group_id, levels = gt_row_grps)) %>%
+    dplyr::arrange(group_id) %>%
+    dplyr::pull(rownum_i)
+
+  # get the actual row id of the data for gt to target
   row_ids <- gt_object[["_data"]] %>%
-    mutate(row_id = row_number()) %>%
-    group_by(!!grp_col) %>%
-    filter({{column}} == do.call(what = fn, args = list({{column}}))) %>%
-    pull(.data$row_id)
+    dplyr::mutate(row_id = dplyr::row_number()) %>%
+    dplyr::slice(grp_vec_ord) %>%
+    ### !!! to evaluate the list of symbols
+    dplyr::group_by(!!!grp_col) %>%
+    dplyr::filter({{column}} == do.call(what = fn, args = list({{column}}))) %>%
+    dplyr::ungroup() %>%
+    dplyr::pull(row_id)
 
   gt_object %>%
     tab_style(